+++ coding/pygments/logtalk.py
@@ -66,6 +66,7 @@ class LogtalkLexer(RegexLexer):
             (r'(current|set)_logtalk_flag(?=[(])', Keyword),
             # Compiling, loading, and library paths
             (r'logtalk_(compile|l(ibrary_path|oad|oad_context))(?=[(])', Keyword),
+            (r'\blogtalk_make\b', Keyword),
             # Database
             (r'(clause|retract(all)?)(?=[(])', Keyword),
             (r'a(bolish|ssert(a|z))(?=[(])', Keyword),